Move combobox into SpecInput
